+++ pynhd/__init__.py
@@ -4,7 +4,7 @@ from pkg_resources import DistributionNotFound, get_distribution
 from .exceptions import InvalidInputType, InvalidInputValue, MissingItems, ZeroMatched
 from .network_tools import prepare_nhdplus, topoogical_sort, vector_accumulation
 from .print_versions import show_versions
-from .pynhd import NLDI, AGRBase, NHDPlusHR, WaterData
+from .pynhd import NLDI, AGRBase, NHDPlusHR, WaterData, nhdplus_attrs
 
 try:
     __version__ = get_distribution(__name__).version